Add ability to check for existence of item in collection for action field list criteria evaluators
