+++ Assets/Scripts/Incidents/IncidentActionFieldCriteria.cs
@@ -33,6 +33,10 @@ namespace Game.Incidents
 			{
 				evaluator = new ActionFieldIntDictionaryEvaluator(propertyName, ContextType);
 			}
+			else if(PrimitiveType == typeof(List<IIncidentContext>))
+			{
+				evaluator = new ActionFieldListContainsEvaluator(propertyName, ContextType);
+			}
 		}
 	}
 }
